// src/store/enums/RoleActionsEnum.ts
export enum RoleActionType {
  GET_ALL_ROLES = "GET_ALL_ROLES",
  GET_ALL_ROLES_SUCCESS = "GET_ALL_ROLES_SUCCESS",
  GET_ALL_ROLES_FAILURE = "GET_ALL_ROLES_FAILURE",
  GET_ROLE = "GET_ROLE",
  GET_ROLE_SUCCESS = "GET_ROLE_SUCCESS",
  GET_ROLE_FAILURE = "GET_ROLE_FAILURE",
  CREATE_ROLE = "CREATE_ROLE",
  CREATE_ROLE_SUCCESS = "CREATE_ROLE_SUCCESS",
  CREATE_ROLE_FAILURE = "CREATE_ROLE_FAILURE",
  UPDATE_ROLE = "UPDATE_ROLE",
  UPDATE_ROLE_SUCCESS = "UPDATE_ROLE_SUCCESS",
  UPDATE_ROLE_FAILURE = "UPDATE_ROLE_FAILURE",
  DELETE_ROLE = "DELETE_ROLE",
  DELETE_ROLE_SUCCESS = "DELETE_ROLE_SUCCESS",
  DELETE_ROLE_FAILURE = "DELETE_ROLE_FAILURE",
  GET_ALL_ACTIVITIES = "GET_ALL_ACTIVITIES",
  GET_ALL_ACTIVITIES_SUCCESS = "GET_ALL_ACTIVITIES_SUCCESS",
  GET_ALL_ACTIVITIES_FAILURE = "GET_ALL_ACTIVITIES_FAILURE",
}